+++ lib/src/interceptors/chopper_interceptor.dart
@@ -1,7 +1,6 @@
-// ignore_for_file: use_if_null_to_convert_nulls_to_bools
-
 import 'dart:async';
 import 'dart:convert';
+import 'dart:developer';
 
 import 'package:chopper/chopper.dart';
 import 'package:chucker_flutter/src/helpers/constants.dart';
@@ -11,20 +10,20 @@ import 'package:chucker_flutter/src/view/helper/chucker_ui_helper.dart';
 import 'package:http/http.dart' as http;
 
 ///[ChuckerChopperInterceptor] adds support for `chucker_flutter` in Chopper
-class ChuckerChopperInterceptor extends ResponseInterceptor {
+class ChuckerChopperInterceptor implements ResponseInterceptor {
   @override
   FutureOr<Response<dynamic>> onResponse(Response<dynamic> response) async {
     final time = DateTime.now();
     await SharedPreferencesManager.getInstance().getSettings();
 
     if (ChuckerFlutter.isDebugMode || ChuckerFlutter.showOnRelease) {
-      await _saveResponse(response, time);
       await ChuckerUiHelper.showNotification(
         method: response.base.request?.method ?? '',
         statusCode: response.statusCode,
         path: response.base.request?.url.path ?? '',
         requestTime: time,
       );
+      await _saveResponse(response, time);
     }
     return response;
   }
@@ -52,7 +51,7 @@ class ChuckerChopperInterceptor extends ResponseInterceptor {
             response.base.request?.url.queryParameters.toString() ??
                 emptyString,
         receiveTimeout: 0,
-        request: {'request': _requestBody(response)},
+        request: _requestBody(response),
         requestSize: 2,
         requestTime: time,
         responseSize: 2,
@@ -63,6 +62,12 @@ class ChuckerChopperInterceptor extends ResponseInterceptor {
         clientLibrary: 'Chopper',
       ),
     );
+
+    final method = response.base.request?.method ?? '';
+    final statusCode = response.statusCode;
+    final path = response.base.request?.url.path ?? '';
+
+    log('ChuckerFlutter: $method:$path($statusCode) saved.');
   }
 
   String _requestType(Response<dynamic> response) {
@@ -82,17 +87,15 @@ class ChuckerChopperInterceptor extends ResponseInterceptor {
     }
 
     if (response.base.request is http.Request) {
-      final request = response.base.request as http.Request?;
-      return request?.body.isNotEmpty == true
-          ? _getRequestBody(request)
-          : emptyString;
+      final request = response.base.request! as http.Request;
+      return request.body.isNotEmpty ? _getRequestBody(request) : emptyString;
     }
     return emptyString;
   }
 
-  dynamic _getRequestBody(http.Request? request) {
+  dynamic _getRequestBody(http.Request request) {
     try {
-      return jsonDecode(request!.body);
+      return jsonDecode(request.body);
       // ignore: empty_catches
     } catch (e) {}
   }